Refactor directivity
