+++ zea/func/__init__.py
@@ -59,6 +59,7 @@ from .ultrasound import (
     dehaze_nuclear_diffusion,
     demodulate,
     demodulate_not_jitable,
+    directivity,
     envelope_detect,
     get_band_pass_filter,
     get_low_pass_iq_filter,
@@ -127,6 +128,7 @@ __all__ = [
     "dehaze_nuclear_diffusion",
     "demodulate",
     "demodulate_not_jitable",
+    "directivity",
     "envelope_detect",
     "get_band_pass_filter",
     "get_low_pass_iq_filter",
